optimized the components directory structure
