+++ components/modal/Modal.tsx
@@ -1,6 +1,7 @@
 import { useCallback } from "react";
 import { AiOutlineClose } from "react-icons/ai";
-import Button from "./Button";
+import Button from "../layout/Button";
+
 
 interface ModalProps {
   isOpen?: boolean;
