+++ components/sidebar/FollowBar.tsx
@@ -1,6 +1,6 @@
 import useUsers from '@/hooks/useUsers';
 
-import Avatar from '../Avatar';
+import Avatar from '../user/Avatar';
 
 const FollowBar = () => {
   const { data: users = [] } = useUsers();
