+++ pages/_app.tsx
@@ -1,14 +1,14 @@
 import Layout from "@/components/Layout";
-import LoginModal from "@/components/modals/LoginModal";
-import RegisterModal from "@/components/modals/RegisterModal";
+import LoginModal from "@/components/modal/LoginModal";
+import RegisterModal from "@/components/modal/RegisterModal";
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
 import { Toaster } from "react-hot-toast";
 import { SessionProvider } from "next-auth/react";
-import EditModal from "@/components/modals/EditModal";
+import EditModal from "@/components/modal/EditModal";
 import Head from "next/head";
 import Favicon from "@/components/Favicon";
-import PostModal from "@/components/modals/PostModal";
+import PostModal from "@/components/modal/PostModal";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
