+++ pages/posts/[postId].tsx
@@ -1,7 +1,7 @@
 import Form from "@/components/Form";
 import Header from "@/components/Header";
-import CommentFeed from "@/components/posts/CommentFeed";
-import PostItem from "@/components/posts/PostItem";
+import CommentFeed from "@/components/post/CommentFeed";
+import PostItem from "@/components/post/PostItem";
 import usePost from "@/hooks/usePost";
 import { useRouter } from "next/router";
 import { ClipLoader } from "react-spinners";
