+++ pages/users/[userId].tsx
@@ -2,9 +2,9 @@ import Header from "@/components/Header";
 import useUser from "@/hooks/useUser";
 import { useRouter } from "next/router";
 import { ClipLoader } from "react-spinners";
-import UserHero from "@/components/users/UserHero";
-import UserBio from "../../components/users/UserBio";
-import PostFeed from "@/components/posts/PostFeed";
+import UserHero from "@/components/user/UserHero";
+import UserBio from "../../components/user/UserBio";
+import PostFeed from "@/components/post/PostFeed";
 
 const UserView = () => {
     const router = useRouter();
